feat: configure monitoring (#117)
## Description
Enabled service and pod monitors in gitlab. For running on k3d-slim-dev
in ci/local we dont have to update anything as the gitlab chart checks
for the monitoring capabilities presence to deploy the service monitor
templates.

Depends on uds-core
[0.21.0](https://github.com/defenseunicorns/uds-core/releases/tag/v0.21.0).

Note: the pepr capability to mutate the service monitors but not pod
monitors so we will need to request that feature or PR uds-core to add
it. Adding the config for the pod monitor does no break anything.

As this wont get tested in CI in this repo (we should consider the
"on-release" test that will exist soon to deploy a full cluster) I will
provide visual proof in the comments. This requires no changes to the
end user to enable.

## Related Issue

Fixes #
[116](https://github.com/defenseunicorns/uds-package-gitlab/issues/116)
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-gitlab/blob/main/CONTRIBUTING.md#developer-workflow)
followed