feat: add configurable bot accounts and gitlab license (#206)
## Description

Add ability to configure bot accounts when deploying gitlab. For each
account it will create the account in gitlab, create a pat for the
account, and store the pat in a secret. If using gitlab premium or
ultimate, it will create service accounts instead of regular user
accounts.

Also added the ability to set a gitlab license in the config chart since
that was needed for local testing on deploy. Tested that locally with a
trial license. Will add screenshot of that test result in comments.

## Related Issue

Fixes #205 
Fixes #207 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-gitlab/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------

Co-authored-by: Wayne Starr <Racer159@users.noreply.github.com>
Release-As: v17.2.7-uds.2