fix: added gitaly cgroups init image to package (#246)
## Description

Downstream we use this image to modify cgroups on gitaly nodes to
improve resiliency. This isn't in the package right now, so we're having
to add it to a local package so it is available. By putting it here in
the original package we ensure the version always matches the rest of
the install, and save others who do similar work time later.

## Type of change

- [X] Bug fix (non-breaking change which fixes an issue)

## Checklist before merging

- [X] Test, docs, adr added or updated as needed
- [X] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-gitlab/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------

Co-authored-by: Rob <robert.mcelvenny@gmail.com>
Co-authored-by: Wayne Starr <Racer159@users.noreply.github.com>
Co-authored-by: Rob McElvenny <robmc@defenseunicorns.com>

Release-As: v17.6.1-uds.1